CXR_ViewCLS/Test/requirements.txt
absl-py==0.8.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
astor==0.8.0
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
bleach==4.1.0
cffi==1.15.1
cryptography==2.1.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.0
h5py==2.10.0
idna==2.6
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.5
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
networkx==2.5.1
nibabel==3.2.2
notebook==6.4.10
numpy==1.17.2
opencv-python==4.8.0.74
opt-einsum==3.1.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.9.2
ptyprocess==0.7.0
pycparser==2.21
pycrypto==2.6.1
pydicom==2.3.1
Pygments==2.14.0
PyGObject==3.26.1
pylibjpeg==1.3.0
pylibjpeg-libjpeg==1.2.0
pylibjpeg-openjpeg==1.1.1
pylibjpeg-rle==1.1.0
pyparsing==3.1.0
pyrsistent==0.18.0
python-apt==1.6.4
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.1.1
pyxdg==0.25
pyzmq==25.1.0
qtconsole==5.2.2
QtPy==2.0.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.2
SimpleITK==1.2.4
six==1.11.0
tensorboard==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.6.4
wrapt==1.11.2
zipp==3.6.0